requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
langchain==0.2.12
langchain-openai==0.1.22
langchain-community==0.2.11
langgraph==0.2.27
openai==1.40.3
pypdf
faiss-cpu==1.8.0.post1
duckduckgo-search==6.2.13
python-dotenv==1.0.1
pydantic==2.8.2
pytest==8.3.2
httpx==0.27.2
tavily-python==0.3.5